Move utilization_duration methods from `AmazonSageMakerBaseExecutor` to training
Utilization and cost tracking is implemented differently for endpoint. By moving it out of AmazonSageMakerBaseExecutor, this difference is now explicit.
